+++ org.jacoco.ant.test/src/org/jacoco/ant/CreateExecFiles.java
@@ -40,7 +40,7 @@ public class CreateExecFiles {
 		out = new FileOutputStream(BASE_LOCATION + "nomatch.exec");
 		ExecutionDataWriter writer = new ExecutionDataWriter(out);
 		writer.visitClassExecution(
-				new ExecutionData(0, "org/jacoco/ant/TestTarget", new int[0]));
+				new ExecutionData(0, "org/jacoco/ant/TestTarget", new long[0]));
 		out.close();
 	}
