+++ org.jacoco.core.test/src/org/jacoco/core/data/ExecutionDataReaderWriterTest.java
@@ -72,7 +72,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test
 	public void testFlush() throws IOException {
-		final int[] flushCalled = new int[] { 0 };
+		final long[] flushCalled = new long[] { 0 };
 		final OutputStream out = new OutputStream() {
 			@Override
 			public void write(int b) throws IOException {
@@ -200,7 +200,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test(expected = RuntimeException.class)
 	public void testSessionInfoIOException() throws IOException {
-		final int[] broken = new int[1];
+		final long[] broken = new long[1];
 		final ExecutionDataWriter writer = createWriter(new OutputStream() {
 			@Override
 			public void write(int b) throws IOException {
@@ -224,7 +224,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test
 	public void testMinClassId() throws IOException {
-		final int[] data = createData(8);
+		final long[] data = createData(8);
 		writer.visitClassExecution(
 				new ExecutionData(Long.MIN_VALUE, "Sample", data));
 		assertFalse(createReaderWithVisitors().read());
@@ -233,7 +233,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test
 	public void testMaxClassId() throws IOException {
-		final int[] data = createData(8);
+		final long[] data = createData(8);
 		writer.visitClassExecution(
 				new ExecutionData(Long.MAX_VALUE, "Sample", data));
 		assertFalse(createReaderWithVisitors().read());
@@ -242,7 +242,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test
 	public void testEmptyClass() throws IOException {
-		final int[] data = createData(0);
+		final long[] data = createData(0);
 		writer.visitClassExecution(new ExecutionData(3, "Sample", data));
 		assertFalse(createReaderWithVisitors().read());
 		assertTrue(store.getContents().isEmpty());
@@ -250,7 +250,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test
 	public void testNoHitClass() throws IOException {
-		final int[] data = new int[] { 0, 0, 0 };
+		final long[] data = new long[] { 0, 0, 0 };
 		writer.visitClassExecution(new ExecutionData(3, "Sample", data));
 		assertFalse(createReaderWithVisitors().read());
 		assertTrue(store.getContents().isEmpty());
@@ -258,7 +258,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test
 	public void testOneClass() throws IOException {
-		final int[] data = createData(15);
+		final long[] data = createData(15);
 		writer.visitClassExecution(new ExecutionData(3, "Sample", data));
 		assertFalse(createReaderWithVisitors().read());
 		assertArrayEquals(data, store.get(3).getProbes());
@@ -266,8 +266,8 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test
 	public void testTwoClasses() throws IOException {
-		final int[] data1 = createData(15);
-		final int[] data2 = createData(185);
+		final long[] data1 = createData(15);
+		final long[] data2 = createData(185);
 		writer.visitClassExecution(new ExecutionData(333, "Sample", data1));
 		writer.visitClassExecution(new ExecutionData(-45, "Sample", data2));
 		assertFalse(createReaderWithVisitors().read());
@@ -277,7 +277,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test
 	public void testBigClass() throws IOException {
-		final int[] data = createData(3599);
+		final long[] data = createData(3599);
 		writer.visitClassExecution(new ExecutionData(123, "Sample", data));
 		assertFalse(createReaderWithVisitors().read());
 		assertArrayEquals(data, store.get(123).getProbes());
@@ -285,7 +285,7 @@ public class ExecutionDataReaderWriterTest {
 
 	@Test(expected = RuntimeException.class)
 	public void testExecutionDataIOException() throws IOException {
-		final int[] broken = new int[1];
+		final long[] broken = new long[1];
 		final ExecutionDataWriter writer = createWriter(new OutputStream() {
 			@Override
 			public void write(int b) throws IOException {
@@ -296,7 +296,7 @@ public class ExecutionDataReaderWriterTest {
 		});
 		broken[0] = 1;
 		writer.visitClassExecution(
-				new ExecutionData(3, "Sample", new int[] { 1 }));
+				new ExecutionData(3, "Sample", new long[] { 1 }));
 	}
 
 	private ExecutionDataReader createReaderWithVisitors() throws IOException {
@@ -310,15 +310,15 @@ public class ExecutionDataReaderWriterTest {
 		return reader;
 	}
 
-	private int[] createData(final int probeCount) {
-		final int[] data = new int[probeCount];
+	private long[] createData(final int probeCount) {
+		final long[] data = new long[probeCount];
 		for (int j = 0; j < data.length; j++) {
-			data[j] = random.nextInt();
+			data[j] = random.nextLong();
 		}
 		return data;
 	}
 
-	private void assertArrayEquals(final int[] expected, final int[] actual) {
+	private void assertArrayEquals(final long[] expected, final long[] actual) {
 		assertTrue(Arrays.equals(expected, actual));
 	}
