+++ org.jacoco.core/src/org/jacoco/core/analysis/ILine.java
@@ -50,5 +50,5 @@ public interface ILine {
 	 *
 	 * @return execution count
 	 */
-	int getExecutionCount();
+	long getExecutionCount();
 }
