+++ org.jacoco.core/src/org/jacoco/core/internal/analysis/ClassAnalyzer.java
@@ -34,7 +34,7 @@ public class ClassAnalyzer extends ClassProbesVisitor
 		implements IFilterContext {
 
 	private final ClassCoverageImpl coverage;
-	private final int[] probes;
+	private final long[] probes;
 	private final StringPool stringPool;
 
 	private final Set<String> classAnnotations = new HashSet<String>();
@@ -55,7 +55,7 @@ public class ClassAnalyzer extends ClassProbesVisitor
 	 * @param stringPool
 	 *            shared pool to minimize the number of {@link String} instances
 	 */
-	public ClassAnalyzer(final ClassCoverageImpl coverage, final int[] probes,
+	public ClassAnalyzer(final ClassCoverageImpl coverage, final long[] probes,
 			final StringPool stringPool) {
 		this.coverage = coverage;
 		this.probes = probes;
